+++ Assets/Scripts/CameraController.cs
@@ -27,6 +27,8 @@ public class CameraController : MonoBehaviour
     public float MouseSensitivityX = 4;
     public float MouseSensitivityY = 4;
     public float MouseSensitivityScroll = 2;
+    public float MobileSensitivity = 12;
+    public float MobileSensitivityScroll = 60;
 
     /// <summary>
     /// This is our max and min zoom values. It should be updated in the inspector.
@@ -54,6 +56,11 @@ public class CameraController : MonoBehaviour
     private bool allowCameraUpdate = false;
 
     /// <summary>
+    /// This is the previous distance between the users fingers on mobile
+    /// </summary>
+    private float prevDistance = 0;
+
+    /// <summary>
     /// In this start function we set our max and min zoom to the negative version of the public version.
     /// </summary>
     void Start()
@@ -68,6 +75,9 @@ public class CameraController : MonoBehaviour
     void Update()
     {
         MouseInput();
+        MobileInput();
+        HandleZoom();
+        MobileZoom();
         if(allowCameraUpdate) UpdateCamera();
     }
 
@@ -90,12 +100,34 @@ public class CameraController : MonoBehaviour
     /// </summary>
     void MouseInput()
     {
-        scroll = Input.GetAxis("Mouse ScrollWheel");
-        if (scroll != 0) {
-            allowCameraUpdate = false;
-            transform.localPosition = new Vector3(transform.localPosition.x, transform.localPosition.y, transform.localPosition.z + scroll * MouseSensitivityScroll);
+        if (Input.GetButton("Fire1"))
+        {
+            yaw = Input.GetAxis("Mouse X");
+            camPitchAmount = Input.GetAxis("Mouse Y") * MouseSensitivityY * (invertLookY ? -1 : 1);
+
+            HandleOrbit(yaw, camPitchAmount);
         }
 
+        LimitOrbit();
+    }
+
+    void HandleZoom()
+    {
+        scroll = Input.GetAxis("Mouse ScrollWheel");
+        //print(scroll);
+        if (scroll != 0)
+        {
+            allowCameraUpdate = false;
+            transform.localPosition = new Vector3(transform.localPosition.x, transform.localPosition.y, transform.localPosition.z + scroll * MouseSensitivityScroll);
+            LimitZoom();
+        }
+    }
+
+    /// <summary>
+    /// This funciton will limit the minimum and maximum zoom.
+    /// </summary>
+    void LimitZoom()
+    {
         if (transform.localPosition.z > minZoom)
         {
             transform.localPosition = new Vector3(transform.localPosition.x, transform.localPosition.y, minZoom);
@@ -105,21 +137,64 @@ public class CameraController : MonoBehaviour
         {
             transform.localPosition = new Vector3(transform.localPosition.x, transform.localPosition.y, maxZoom);
         }
+    }
 
-        if (Input.GetButton("Fire1"))
+    void MobileInput()
+    {
+        if(Input.touchCount == 1 && Input.GetTouch(0).phase == TouchPhase.Moved)
         {
-            allowCameraUpdate = false;
-            float prevY = camPitch.rotation.eulerAngles.y;
-            float prevZ = camPitch.rotation.eulerAngles.z;
-
-            yaw = Input.GetAxis("Mouse X");
-            camPitchAmount = Input.GetAxis("Mouse Y") * MouseSensitivityY * (invertLookY ? -1 : 1);
-            camPitchAmount = Mathf.Clamp(camPitchAmount, -10, 80);
-
-            camYaw.Rotate(0, yaw * MouseSensitivityX, 0);
-            camPitch.Rotate(camPitchAmount, 0, 0);
+            yaw = Input.GetTouch(0).deltaPosition.x / MobileSensitivity * -1;
+            camPitchAmount = Input.GetTouch(0).deltaPosition.y / MobileSensitivity * MouseSensitivityY * (invertLookY ? -1 : 1) * -1;
+            
+            HandleOrbit(yaw, camPitchAmount);
         }
 
+        LimitOrbit();
+    }
+
+
+    void MobileZoom()
+    {
+        if (Input.touchCount == 2 && Input.GetTouch(0).phase == TouchPhase.Moved)
+        {
+            Vector2 point1 = Input.GetTouch(0).position;
+            Vector2 point2 = Input.GetTouch(1).position;
+
+            //Calculate the difference between the positions
+            float difference = Vector2.Distance(point1, point2);
+
+            if(prevDistance == 0)
+            {
+                prevDistance = difference;
+                return;
+            }else
+            {
+                float zoomAmount = difference - prevDistance;
+                zoomAmount = zoomAmount / MobileSensitivityScroll;
+                allowCameraUpdate = false;
+                transform.localPosition = new Vector3(transform.localPosition.x, transform.localPosition.y, transform.localPosition.z + zoomAmount);
+                LimitZoom();
+            }
+
+            prevDistance = difference;
+        }else
+        {
+            prevDistance = 0;
+        }
+    }
+
+    void HandleOrbit(float yaw, float pitchAmount)
+    {
+        allowCameraUpdate = false;
+
+        pitchAmount = Mathf.Clamp(pitchAmount, -10, 80);
+
+        camYaw.Rotate(0, yaw * MouseSensitivityX, 0);
+        camPitch.Rotate(pitchAmount, 0, 0);
+    }
+
+    void LimitOrbit()
+    {
         //This limits the pitch to a minimum preset angel
         if (camPitch.rotation.eulerAngles.x < 1 || camPitch.rotation.eulerAngles.x > 300)
         {
@@ -137,8 +212,6 @@ public class CameraController : MonoBehaviour
         {
             camPitch.eulerAngles = new Vector3(camPitch.rotation.eulerAngles.x, camPitch.rotation.eulerAngles.y, 0);
         }
-
-        //print("Euler Angles: " + camPitch.eulerAngles);
     }
 
     /// <summary>